requirements.txt
altair==5.2.0
annotated-types==0.6.0
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
click==8.1.7
dask-glm==0.3.0
dask-ml==2023.3.24
dask-xgboost==0.2.0
distro==1.9.0
et-xmlfile==1.1.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
importlib-metadata==7.0.1
joblib==1.3.2
llvmlite==0.41.0
markdown-it-py==3.0.0
mdurl==0.1.2
multipledispatch==1.0.0
numba==0.58.0
numpy==1.26.4
openai==1.13.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pipx==1.2.1
playwright==1.42.0
protobuf==4.25.3
pyarrow==15.0.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pyee==11.0.1
python-dotenv==1.0.1
rich==13.7.1
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.18.1
smmap==5.0.1
sortedcontainers==2.4.0
sparse==0.14.0
streamlit==1.31.1
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.1
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
tzdata==2024.1
tzlocal==5.2
undetected-chromedriver==3.5.5
userpath==1.9.1
validators==0.22.0
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
wget==3.2
wsproto==1.2.0
xgboost==0.90
zipp==3.17.0
